refactor: decompose RunHarnessOptimizationCommandHandler (647→281+160+162) into partials
Split by responsibility: orchestration, evaluation/scoring, proposal generation.
