+++ src/Content/Application/Application.Core/CQRS/MetaHarness/RunHarnessOptimizationCommandHandler.cs
@@ -1,5 +1,3 @@
-using System.Security.Cryptography;
-using System.Text;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using Application.AI.Common.Exceptions;
@@ -31,7 +29,7 @@ namespace Application.Core.CQRS.MetaHarness;
 /// <see cref="OperationCanceledException"/> always propagates to the caller.
 /// </para>
 /// </remarks>
-public sealed class RunHarnessOptimizationCommandHandler
+public sealed partial class RunHarnessOptimizationCommandHandler
     : IRequestHandler<RunHarnessOptimizationCommand, OptimizationResult>
 {
     private readonly IHarnessProposer _proposer;
@@ -295,336 +293,6 @@ public sealed class RunHarnessOptimizationCommandHandler
             ProposedChangesPath = bestCandidate is not null ? proposedDir : string.Empty,
             EarlyStopReason = earlyStopReason,
         };
-    }
-
-    private static async Task<IReadOnlyList<EvalTask>> LoadEvalTasksAsync(
-        string path, CancellationToken ct)
-    {
-        if (!Directory.Exists(path))
-            return [];
-
-        var tasks = new List<EvalTask>();
-        foreach (var file in Directory.EnumerateFiles(path, "*.json"))
-        {
-            try
-            {
-                var json = await File.ReadAllTextAsync(file, ct);
-                var task = JsonSerializer.Deserialize<EvalTask>(json, JsonOptions);
-                if (task is not null)
-                    tasks.Add(task);
-            }
-            catch (Exception ex) when (ex is JsonException or IOException)
-            {
-                // Skip individual malformed/unreadable task files; directory-level errors propagate
-            }
-        }
-        return tasks;
-    }
-
-    private static async Task<RunManifest> LoadOrCreateRunManifest(string runDir, Guid optimizationRunId)
-    {
-        var path = Path.Combine(runDir, "run_manifest.json");
-        if (File.Exists(path))
-        {
-            try
-            {
-                var json = await File.ReadAllTextAsync(path);
-                var existing = JsonSerializer.Deserialize<RunManifest>(json, JsonOptions);
-                if (existing is not null)
-                    return existing;
-            }
-            catch (Exception ex) when (ex is JsonException or IOException)
-            {
-                // Corrupt manifest — recreate from scratch
-            }
-        }
-
-        return new RunManifest
-        {
-            OptimizationRunId = optimizationRunId,
-            LastCompletedIteration = 0,
-            BestCandidateId = null,
-            StartedAt = DateTimeOffset.UtcNow,
-            WriteCompleted = false,
-        };
-    }
-
-    private static void UpdateRunManifest(
-        string runDir,
-        int lastCompletedIteration,
-        Guid? bestCandidateId,
-        Guid optimizationRunId,
-        DateTimeOffset startedAt)
-    {
-        var updated = new RunManifest
-        {
-            OptimizationRunId = optimizationRunId,
-            LastCompletedIteration = lastCompletedIteration,
-            BestCandidateId = bestCandidateId,
-            StartedAt = startedAt,
-            WriteCompleted = true,
-        };
-        var json = JsonSerializer.Serialize(updated, JsonOptions);
-        var path = Path.Combine(runDir, "run_manifest.json");
-        var tmp = path + ".tmp";
-        File.WriteAllText(tmp, json);
-        File.Move(tmp, path, overwrite: true);
-    }
-
-    /// <summary>
-    /// Returns true when <paramref name="candidate"/> should replace <paramref name="currentBest"/>.
-    /// </summary>
-    /// <remarks>
-    /// Tie-breaking order:
-    /// <list type="number">
-    ///   <item>Pass rate exceeds current best by more than <paramref name="threshold"/> → clear improvement.</item>
-    ///   <item>Pass rate within threshold (tie) → lower token cost wins.</item>
-    ///   <item>Pass rate and cost both tied → lower iteration wins (keep earlier candidate).</item>
-    /// </list>
-    /// </remarks>
-    private static bool IsBetter(
-        HarnessCandidate candidate,
-        HarnessCandidate? currentBest,
-        double threshold)
-    {
-        if (currentBest is null)
-            return true;
-
-        var delta = candidate.BestScore.GetValueOrDefault()
-                    - currentBest.BestScore.GetValueOrDefault();
-
-        if (delta > threshold)
-            return true;
-
-        if (delta < -threshold)
-            return false;
-
-        // Within threshold — tie-break by token cost
-        var costDelta = candidate.TokenCost.GetValueOrDefault()
-                        - currentBest.TokenCost.GetValueOrDefault();
-        if (costDelta < 0) return true;
-        if (costDelta > 0) return false;
-
-        // Tie on cost — prefer earlier iteration
-        return candidate.Iteration < currentBest.Iteration;
-    }
-
-    private static void WriteSnapshotFiles(string runDir, HarnessCandidate candidate)
-    {
-        var snapshotDir = Path.Combine(
-            runDir, "candidates", candidate.CandidateId.ToString(), "snapshot");
-        Directory.CreateDirectory(snapshotDir);
-
-        foreach (var (relativePath, content) in candidate.Snapshot.SkillFileSnapshots)
-        {
-            var filePath = SafeResolvePath(snapshotDir, relativePath);
-            var dir = Path.GetDirectoryName(filePath);
-            if (dir is not null)
-                Directory.CreateDirectory(dir);
-            File.WriteAllText(filePath, content);
-        }
-    }
-
-    private static void WriteProposedSnapshot(string proposedDir, HarnessCandidate? best)
-    {
-        if (best is null)
-            return;
-
-        if (Directory.Exists(proposedDir))
-            Directory.Delete(proposedDir, recursive: true);
-        Directory.CreateDirectory(proposedDir);
-
-        foreach (var (relativePath, content) in best.Snapshot.SkillFileSnapshots)
-        {
-            var filePath = SafeResolvePath(proposedDir, relativePath);
-            var dir = Path.GetDirectoryName(filePath);
-            if (dir is not null)
-                Directory.CreateDirectory(dir);
-            File.WriteAllText(filePath, content);
-        }
-    }
-
-    /// <summary>
-    /// Resolves <paramref name="relativePath"/> under <paramref name="baseDir"/> and asserts
-    /// the result stays within <paramref name="baseDir"/>. Throws on path traversal attempts.
-    /// </summary>
-    private static string SafeResolvePath(string baseDir, string relativePath)
-    {
-        var resolved = Path.GetFullPath(Path.Combine(baseDir, relativePath));
-        var rootWithSep = baseDir.TrimEnd(Path.DirectorySeparatorChar) + Path.DirectorySeparatorChar;
-        if (!resolved.StartsWith(rootWithSep, StringComparison.OrdinalIgnoreCase)
-            && !string.Equals(resolved, baseDir, StringComparison.OrdinalIgnoreCase))
-        {
-            throw new InvalidOperationException(
-                $"Proposed skill path '{relativePath}' resolves outside the snapshot directory.");
-        }
-        return resolved;
-    }
-
-    private async Task WriteSummaryMarkdownAsync(
-        string runDir, Guid optimizationRunId, CancellationToken ct)
-    {
-        var candidates = await _candidateRepository.ListAsync(optimizationRunId, ct);
-        var sb = new StringBuilder();
-        sb.AppendLine("# Optimization Run Summary");
-        sb.AppendLine();
-        sb.AppendLine("| Iteration | CandidateId | PassRate | TokenCost | Status |");
-        sb.AppendLine("|-----------|-------------|----------|-----------|--------|");
-        foreach (var c in candidates.OrderBy(x => x.Iteration))
-        {
-            sb.AppendLine(
-                $"| {c.Iteration} | {c.CandidateId:N} " +
-                $"| {c.BestScore?.ToString("P2") ?? "-"} " +
-                $"| {c.TokenCost?.ToString() ?? "-"} " +
-                $"| {c.Status} |");
-        }
-        await File.WriteAllTextAsync(Path.Combine(runDir, "summary.md"), sb.ToString(), ct);
-    }
-
-    private void EnforceRetentionPolicy(
-        int maxRunsToKeep, string traceDirectoryRoot, Guid currentRunId)
-    {
-        if (maxRunsToKeep <= 0)
-            return;
-
-        var optimizationsDir = Path.Combine(traceDirectoryRoot, "optimizations");
-        if (!Directory.Exists(optimizationsDir))
-            return;
-
-        var others = Directory.GetDirectories(optimizationsDir)
-            .Where(d => Guid.TryParse(Path.GetFileName(d), out _)
-                        && !string.Equals(
-                            Path.GetFileName(d), currentRunId.ToString(),
-                            StringComparison.OrdinalIgnoreCase))
-            .Select(d => new DirectoryInfo(d))
-            .OrderBy(d => d.CreationTimeUtc)
-            .ToList();
-
-        var excess = others.Count - (maxRunsToKeep - 1);
-        for (var i = 0; i < excess; i++)
-        {
-            try
-            {
-                Directory.Delete(others[i].FullName, recursive: true);
-                _logger.LogInformation(
-                    "Retention policy: deleted old optimization run '{Dir}'",
-                    others[i].Name);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogWarning(
-                    ex, "Retention policy: failed to delete '{Dir}'", others[i].FullName);
-            }
-        }
-    }
-
-    private async Task<HarnessCandidate> ResolveSeedCandidateAsync(
-        RunHarnessOptimizationCommand command,
-        MetaHarnessConfig config,
-        CancellationToken ct)
-    {
-        if (command.SeedCandidateId.HasValue)
-        {
-            var existing = await _candidateRepository.GetAsync(command.SeedCandidateId.Value, ct);
-            if (existing is null)
-                throw new InvalidOperationException(
-                    $"Seed candidate '{command.SeedCandidateId.Value}' not found in repository.");
-            return existing;
-        }
-
-        var snapshot = await _snapshotBuilder.BuildAsync(
-            config.SeedCandidatePath,
-            systemPrompt: string.Empty,
-            configValues: new Dictionary<string, string>(),
-            ct);
-
-        var seed = new HarnessCandidate
-        {
-            CandidateId = Guid.NewGuid(),
-            OptimizationRunId = command.OptimizationRunId,
-            ParentCandidateId = null,
-            Iteration = 0,
-            CreatedAt = DateTimeOffset.UtcNow,
-            Snapshot = snapshot,
-            Status = HarnessCandidateStatus.Proposed,
-        };
-        await _candidateRepository.SaveAsync(seed, ct);
-        return seed;
-    }
-
-    private static HarnessSnapshot BuildSnapshotFromProposal(
-        HarnessSnapshot current, HarnessProposal proposal)
-    {
-        var skillFiles = new Dictionary<string, string>(current.SkillFileSnapshots);
-        foreach (var (path, content) in proposal.ProposedSkillChanges)
-            skillFiles[path] = content;
-
-        var configSnapshot = new Dictionary<string, string>(current.ConfigSnapshot);
-        foreach (var (key, value) in proposal.ProposedConfigChanges)
-            configSnapshot[key] = value;
-
-        var systemPrompt = proposal.ProposedSystemPromptChange ?? current.SystemPromptSnapshot;
-
-        var manifest = skillFiles
-            .Select(kvp => new SnapshotEntry(kvp.Key, ComputeSha256(kvp.Value)))
-            .ToList();
-
-        return new HarnessSnapshot
-        {
-            SkillFileSnapshots = skillFiles,
-            SystemPromptSnapshot = systemPrompt,
-            ConfigSnapshot = configSnapshot,
-            SnapshotManifest = manifest,
-        };
-    }
-
-    private static string ComputeSha256(string content)
-    {
-        var bytes = SHA256.HashData(Encoding.UTF8.GetBytes(content));
-        return Convert.ToHexStringLower(bytes);
-    }
-
-    private static async Task<string?> ReadLearningsFileAsync(string runDir)
-    {
-        var path = Path.Combine(runDir, "learnings.md");
-        if (!File.Exists(path))
-            return null;
-        var content = await File.ReadAllTextAsync(path);
-        return string.IsNullOrWhiteSpace(content) ? null : content;
-    }
-
-    private static async Task AppendLearningsAsync(string runDir, string entry)
-    {
-        var path = Path.Combine(runDir, "learnings.md");
-        await File.AppendAllTextAsync(path, entry);
-    }
-
-    private static string BuildLearningsEntry(
-        int iteration,
-        HarnessProposal proposal,
-        EvaluationResult evalResult,
-        bool acceptedAsBest)
-    {
-        var sb = new StringBuilder();
-        sb.AppendLine($"## Iteration {iteration}");
-        sb.AppendLine($"- **Status**: {(acceptedAsBest ? "ACCEPTED as new best" : "Not accepted")}");
-        sb.AppendLine($"- **Pass rate**: {evalResult.PassRate:P2}");
-        sb.AppendLine($"- **Token cost**: {evalResult.TotalTokenCost}");
-        sb.AppendLine($"- **Reasoning**: {proposal.Reasoning}");
-        if (!string.IsNullOrWhiteSpace(proposal.Learnings))
-            sb.AppendLine($"- **Proposer observations**: {proposal.Learnings}");
-        sb.AppendLine();
-        return sb.ToString();
-    }
-
-    private static string BuildFailedLearningsEntry(int iteration, string failureReason)
-    {
-        var sb = new StringBuilder();
-        sb.AppendLine($"## Iteration {iteration}");
-        sb.AppendLine("- **Status**: FAILED");
-        sb.AppendLine($"- **Failure**: {failureReason}");
-        sb.AppendLine();
-        return sb.ToString();
     }
 
     private sealed record RunManifest
